requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.26.4
azure-search-documents==11.3.0
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dataclasses-json==0.5.7
faiss-cpu==1.7.4
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
isodate==0.6.1
langchain==0.0.165
marshmallow==3.19.0
marshmallow-enum==1.5.1
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
yarl==1.9.2
